# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "operationId": "DocumentsController_findAll",
        "summary": "Lista documenti",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["draft", "generated", "published", "archived"],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "summary": "Crea documento da template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "CreateDocumentDto",
                "required": ["name", "templateId"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Capitolato Beta"
                  },
                  "templateId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_findOne",
        "summary": "Dettaglio documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "404": {
            "description": "Documento non trovato"
          }
        },
        "tags": ["documents"]
      },
      "put": {
        "operationId": "DocumentsController_update",
        "summary": "Aggiorna contenuto/fieldValues documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "UpdateDocumentDto",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "fieldValues": {
                    "type": "object",
                    "description": "Mappa chiave-valore dei campi template",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "example": {
                      "titolo": "Contratto 2026",
                      "cliente": "Mario Rossi",
                      "importo": "1000"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      },
      "delete": {
        "operationId": "DocumentsController_remove",
        "summary": "Elimina documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/generate-pdf": {
      "post": {
        "operationId": "DocumentsController_generatePdf",
        "summary": "Accoda generazione PDF asincrona",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "202": {
            "description": "Job PDF accodato"
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/pdf/latest": {
      "get": {
        "operationId": "DocumentsController_latestPdf",
        "summary": "Scarica l'ultimo PDF completato",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "File PDF scaricato",
            "content": {
              "application/pdf": {}
            }
          },
          "404": {
            "description": "Nessun PDF disponibile"
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/preview-pdf": {
      "get": {
        "operationId": "DocumentsController_previewPdf",
        "summary": "Anteprima PDF temporanea (non salvata)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/versions": {
      "get": {
        "operationId": "DocumentsController_getVersions",
        "summary": "Cronologia versioni documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/versions/{version}": {
      "get": {
        "operationId": "DocumentsController_getVersionContent",
        "summary": "Contenuto di una versione specifica del documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "description": "Numero versione",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/restore/{version}": {
      "post": {
        "operationId": "DocumentsController_restoreVersion",
        "summary": "Ripristina una versione precedente del documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "description": "Versione da ripristinare",
            "schema": {}
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/export-md": {
      "get": {
        "operationId": "DocumentsController_exportMd",
        "summary": "Esporta documento come file .md",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/documents/{id}/audit": {
      "get": {
        "operationId": "DocumentsController_getAudit",
        "summary": "Audit log del documento",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID documento",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["documents"]
      }
    },
    "/api/pdf/templates/{templateId}/validate": {
      "post": {
        "operationId": "PdfController_validateTemplate",
        "summary": "Valida disponibilità template per generazione PDF",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "description": "UUID template",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Stato disponibilità template"
          },
          "404": {
            "description": "Template non trovato"
          }
        },
        "tags": ["pdf"]
      }
    },
...
  },
  "info": {
    "title": "MAC Documents API",
    "description": "API per la generazione documentale basata su template Markdown.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}